Add files via upload
YAYYYYY
got my program to accept and evaluate and convert, but I need to change the flow of the program to actually print the total hex value of the string. BUT the total hex value is successfully calculated and stored in one of the registers. Very tired, don't feel like fixing it right now.